requirements.txt
boto3
pika
Pillow